[EI-260] [web] refactor: Add MetadataSharePopover component for sharing indicator board metadata
